requirements/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements/requirements.txt requirements/requirements.in
#
attrs==18.1.0             # via pytest
beautifulsoup4==4.6.0
certifi==2018.4.16        # via requests
chardet==3.0.4            # via requests
click==6.7                # via pip-tools
codecov==2.0.15
coverage==4.5.1
django-coverage-plugin==1.5.0
django-debug-toolbar==1.9.1
django-leaflet==0.23.0
django-model-utils==3.1.1
django-webtest==1.9.3
django==2.0.5
djangorestframework==3.8.2
factory-boy==2.11.1
faker==0.8.15             # via factory-boy
first==2.0.1              # via pip-tools
idna==2.6                 # via requests
more-itertools==4.1.0     # via pytest
pip-tools==2.0.2
pkginfo==1.4.2            # via twine
pluggy==0.6.0             # via pytest, tox
psycopg2-binary==2.7.4
psycopg2==2.7.4
py==1.5.3                 # via pytest, tox
pygeocoder==1.2.5
pytest-cov==2.5.1
pytest-django==3.2.1
pytest-runner==4.2
pytest-sugar==0.9.1
pytest==3.5.1
python-dateutil==2.7.2    # via faker
pytz==2018.4              # via django
requests-toolbelt==0.8.0  # via twine
requests==2.18.4          # via codecov, pygeocoder, requests-toolbelt, twine
six==1.11.0               # via django-coverage-plugin, faker, more-itertools, pip-tools, pytest, python-dateutil, tox, webtest
sqlparse==0.2.4           # via django-debug-toolbar
termcolor==1.1.0          # via pytest-sugar
text-unidecode==1.2       # via faker
tox==3.0.0
tqdm==4.23.2              # via twine
twine==1.11.0
urllib3==1.22             # via requests
virtualenv==15.2.0
waitress==1.1.0           # via webtest
webob==1.8.1              # via webtest
webtest==2.0.29           # via django-webtest
wheel==0.31.0

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
